Remove sbt.syntax
