+++ MIGRATION.md
@@ -1,7 +1,7 @@
 Migration notes
 ===============
 
-- Build definition is based on Scala 2.11.8
+- Build definition is based on Scala 2.12.1
 - Build.scala style builds are gone. Use multi-project `build.sbt`.
 - `Project(...)` constructor is restricted down to two parameters. Use `project` instead.
 - `sbt.Plugin` is also gone. Use auto plugins.
@@ -13,10 +13,3 @@ Migration notes
 - Renames early command feature from `--<command>` to `early(<command>)`.
 - Log options `-error`, `-warn`, `-info`, `-debug` are added as shorthand for `"early(error)"` etc.
 - `sbt.Process` and `sbt.ProcessExtra` are gone. Use `scala.sys.process` instead.
-
-#### Additional import required
-
-Implicit conversions are moved to `sbt.syntax`. Add the following imports to auto plugins
-or `project/*.scala`.
-
-    import sbt._, syntax._, Keys._
