+++ sbt/src/sbt-test/actions/aggregate/project/Marker.scala
@@ -1,4 +1,4 @@
-import sbt._, syntax._, Keys._
+import sbt._, Keys._
 import Def.Initialize
 
 object Marker extends AutoPlugin {
