+++ sbt/src/sbt-test/project/auto-import/project/P.scala
@@ -1,4 +1,4 @@
-import sbt._, syntax._
+import sbt._
 
 object Q extends AutoPlugin {
   override val requires = plugins.JvmPlugin
