+++ sbt/src/sbt-test/project/binary-plugin/changes/define/A.scala
@@ -1,6 +1,6 @@
 package sbttest // you need package http://stackoverflow.com/questions/9822008/
 
-import sbt._, syntax._, Keys._
+import sbt._, Keys._
 
 object C extends AutoPlugin {
   object autoImport {
