+++ sbt/src/sbt-test/project/session-update-from-cmd/project/Common.scala
@@ -1,4 +1,4 @@
-import sbt._, syntax._, Keys._
+import sbt._, Keys._
 
 import sbt.internal.SessionSettings
